Add I/O to sets

This merge request enables the capability to read/write train/test sets.

See merge request !4

+++ RBniCS/scm.py
@@ -73,6 +73,8 @@ class SCM(ParametrizedProblem):
         self.truth_A__condensed_for_maximum_eigenvalue = ()
         self.S__condensed = ()
         # 9. I/O
+        self.xi_train_folder = "xi_train__scm/"
+        self.xi_test_folder = "xi_test__scm/"
         self.snap_folder = "snapshots__scm/"
         self.basis_folder = "basis__scm/"
         self.dual_folder = "dual__scm/" # never used
@@ -89,31 +91,24 @@ class SCM(ParametrizedProblem):
     #  @{
     
     ## OFFLINE: set the elements in the training set \xi_train. Overridden to resize alpha_LB_on_xi_train
-    def setxi_train(self, ntrain, sampling="random"):    
-        # Try to import from file
+    ##          Note that the default value of enable_import has been changed here to True
+    def setxi_train(self, ntrain, enable_import=True, sampling="random"):
+        if not enable_import:
+        	sys.exit("SCM will not work without import.")
+        # Save the flag if can import from file
         import_successful = False
-        if os.path.exists(self.red_matrices_folder + "xi_train.npy"):
-            xi_train = np.load(self.red_matrices_folder + "xi_train.npy")
+        if os.path.exists(self.xi_train_folder + "xi_train.npy"):
+            xi_train = np.load(self.xi_train_folder + "xi_train.npy")
             import_successful = (len(np.asarray(xi_train)) == ntrain)
-        # If the imported xi_train has the correct dimensions, use it,
-        # to avoid problems when the offline stage is being commented out,
-        # otherwise generate a new one.
+        # Call parent
+        ParametrizedProblem.setxi_train(self, ntrain, enable_import, sampling)
+        # If xi_train was not imported, be safe and remove the previous folder, so that
+        # if the user overwrites the training set but forgets to run the offline
+        # phase he/she will get an error
         if import_successful == False:
-            ParametrizedProblem.setxi_train(self, ntrain, sampling)
-            # Since the online evaluation depends also on the training set, we need
-            # to save it to file. Be safe and remove the previous folder, so that
-            # if the user overwrites the training set but forgets to run the offline
-            # phase he/she will get an error
             if os.path.exists(self.red_matrices_folder):
                 shutil.rmtree(self.red_matrices_folder)
             os.makedirs(self.red_matrices_folder)
-            # Save the training set to file
-            np.save(self.red_matrices_folder + "xi_train", self.xi_train)
-        else:
-            # Convert to an array of tuples
-            self.xi_train = list()
-            for i in range(len(np.asarray(xi_train))):
-                self.xi_train.append(tuple(xi_train[i, :]))
         # Properly resize related structures
         self.alpha_LB_on_xi_train = np.zeros([ntrain])
         self.complement_C_J = range(ntrain)
